tests updated to reflect changes made to ExtendedLoadableComponent
Two tests are failing. These are the tests of pages that actually are
element containers on the products page: breadcrump and user panel. Its
an obvious result of moving the PageFactory.initElements(driver, this).
This is a sign:) to refactor such 'sub' pages to actual element
containers. This will be a next task)